refactor argocd into generic and deploy classes
