+++ src/ploigos_step_runner/step_implementers/deploy/__init__.py
@@ -1,4 +1,5 @@
 """`StepImplementers` for the `deploy` step.
 """
 
+from ploigos_step_runner.step_implementers.deploy.argocd_deploy import ArgoCDDeploy
 from ploigos_step_runner.step_implementers.deploy.argocd import ArgoCD
